consolidate ProjectList components
